* Fix for Dark Sky response error for offset value in certain timezones * Fix for TTS to use the device volume instead of adjusting it

+++ app/src/main/java/com/thanksmister/iot/mqtt/alarmpanel/network/model/DarkSkyResponse.java
@@ -35,7 +35,7 @@ public class DarkSkyResponse {
     private String timezone;
     @SerializedName("offset")
     @Expose
-    private Integer offset;
+    private Double offset;
     @SerializedName("currently")
     @Expose
     private Currently currently;
@@ -67,11 +67,11 @@ public class DarkSkyResponse {
         this.timezone = timezone;
     }
 
-    public Integer getOffset() {
+    public Double getOffset() {
         return offset;
     }
 
-    public void setOffset(Integer offset) {
+    public void setOffset(Double offset) {
         this.offset = offset;
     }
